pyproject.toml
[project]
name = "mkcollage"
version = "0.1.0"
description = "A simple grid collage maker for creating image collages from a folder"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pillow>=10.0.0",
]

[project.scripts]
mkcollage = "mkcollage.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mkcollage"]

